@@ -49,13 +49,13 @@
     "@sentry-internal/react-inspector": "6.0.1-4",
     "@sentry-internal/rrweb": "1.106.0",
     "@sentry-internal/rrweb-player": "1.106.0",
-    "@sentry/integrations": "^7.45.0",
-    "@sentry/node": "^7.45.0",
+    "@sentry/integrations": "^7.46.0",
+    "@sentry/node": "^7.46.0",
     "@sentry/profiling-node": "^0.3.0",
-    "@sentry/react": "^7.45.0",
+    "@sentry/react": "^7.46.0",
     "@sentry/release-parser": "^1.3.1",
     "@sentry/tracing": "^7.45.0",
-    "@sentry/utils": "^7.45.0",
+    "@sentry/utils": "^7.46.0",
     "@tanstack/react-query": "^4.22.4",
     "@types/color": "^3.0.3",
     "@types/crypto-js": "^4.1.1",
@@ -160,7 +160,7 @@
     "typescript": "^4.9.4",
     "u2f-api": "1.0.10",
     "url-loader": "^4.1.0",
-    "webpack": "5.75.0",
+    "webpack": "5.77.0",
     "webpack-cli": "5.0.1",
     "webpack-remove-empty-scripts": "^1.0.1",
     "zxcvbn": "^4.4.2"
