@@ -1,21 +1,17 @@
-import {Component} from 'react';
+import {useEffect, useState} from 'react';
 import styled from '@emotion/styled';
 
 import robotBackground from 'sentry-images/spot/sentry-robot.png';
 
-import {Client} from 'sentry/api';
 import {Button} from 'sentry/components/button';
 import Link from 'sentry/components/links/link';
 import {t, tct} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import {Organization, Project} from 'sentry/types';
-import {defined} from 'sentry/utils';
-import withApi from 'sentry/utils/withApi';
+import useApi from 'sentry/utils/useApi';
 import CreateSampleEventButton from 'sentry/views/onboarding/createSampleEventButton';
 
 type Props = {
-  api: Client;
-  gradient: boolean;
   org: Organization;
   project?: Project;
   /**
@@ -27,126 +23,103 @@ type Props = {
   sampleIssueId?: string;
 };
 
-type State = {
-  error: boolean;
-  loading: boolean;
-  sampleIssueId?: string;
-};
+function ErrorRobot({org, project, sampleIssueId: sampleIssueIdProp}: Props) {
+  const api = useApi();
 
-class ErrorRobot extends Component<Props, State> {
-  state: State = {
-    error: false,
-    loading: false,
-    sampleIssueId: this.props.sampleIssueId,
-  };
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState<boolean | string>(false);
+  const [sampleIssueId, setSampleIssueId] = useState(sampleIssueIdProp);
 
-  componentDidMount() {
-    this.fetchData();
-  }
+  useEffect(() => {
+    async function loadSampleData() {
+      if (!project) {
+        return;
+      }
 
-  async fetchData() {
-    const {org, project} = this.props;
-    const {sampleIssueId} = this.state;
+      if (sampleIssueIdProp !== undefined) {
+        return;
+      }
 
-    if (!project) {
-      return;
+      setLoading(true);
+
+      try {
+        const data = await api.requestPromise(
+          `/projects/${org.slug}/${project.slug}/issues/`,
+          {
+            method: 'GET',
+            data: {limit: 1},
+          }
+        );
+        setSampleIssueId((data.length > 0 && data[0].id) || '');
+      } catch (err) {
+        setError(err?.responseJSON?.detail ?? true);
+      }
     }
 
-    if (defined(sampleIssueId)) {
-      return;
-    }
+    loadSampleData();
+  }, [api, org, project, sampleIssueIdProp]);
 
-    const url = `/projects/${org.slug}/${project.slug}/issues/`;
-
-    this.setState({loading: true});
-
-    try {
-      const data = await this.props.api.requestPromise(url, {
-        method: 'GET',
-        data: {limit: 1},
-      });
-      this.setState({sampleIssueId: (data.length > 0 && data[0].id) || ''});
-    } catch (err) {
-      const error = err?.responseJSON?.detail ?? true;
-      this.setState({error});
-    }
-
-    this.setState({loading: false});
-  }
-
-  render() {
-    const {loading, error, sampleIssueId} = this.state;
-    const {org, project, gradient} = this.props;
-
-    const sampleLink =
-      project && (loading || error ? null : sampleIssueId) ? (
-        <p>
-          <Link to={`/${org.slug}/${project.slug}/issues/${sampleIssueId}/?sample`}>
-            {t('Or see your sample event')}
-          </Link>
-        </p>
-      ) : (
-        <p>
-          <CreateSampleEventButton
-            priority="link"
-            project={project}
-            source="issues_list"
-            disabled={!project}
-            title={!project ? t('Select a project to create a sample event') : undefined}
-          >
-            {t('Create a sample event')}
-          </CreateSampleEventButton>
-        </p>
-      );
-
-    return (
-      <ErrorRobotWrapper
-        data-test-id="awaiting-events"
-        className="awaiting-events"
-        gradient={gradient}
-      >
-        <Robot aria-hidden>
-          <Eye />
-        </Robot>
-        <MessageContainer>
-          <h3>{t('Waiting for events…')}</h3>
-          <p>
-            {tct(
-              'Our error robot is waiting to [strike:devour] receive your first event.',
-              {
-                strike: <Strikethrough />,
-              }
-            )}
-          </p>
-          <p>
-            {project && (
-              <Button
-                data-test-id="install-instructions"
-                priority="primary"
-                to={`/${org.slug}/${project.slug}/getting-started/${
-                  project.platform || ''
-                }`}
-              >
-                {t('Installation Instructions')}
-              </Button>
-            )}
-          </p>
-          {sampleLink}
-        </MessageContainer>
-      </ErrorRobotWrapper>
+  const sampleLink =
+    project && (loading || error ? null : sampleIssueId) ? (
+      <p>
+        <Link to={`/${org.slug}/${project.slug}/issues/${sampleIssueId}/?sample`}>
+          {t('Or see your sample event')}
+        </Link>
+      </p>
+    ) : (
+      <p>
+        <CreateSampleEventButton
+          priority="link"
+          project={project}
+          source="issues_list"
+          disabled={!project}
+          title={!project ? t('Select a project to create a sample event') : undefined}
+        >
+          {t('Create a sample event')}
+        </CreateSampleEventButton>
+      </p>
     );
-  }
+
+  return (
+    <ErrorRobotWrapper data-test-id="awaiting-events" className="awaiting-events">
+      <Robot aria-hidden>
+        <Eye />
+      </Robot>
+      <MessageContainer>
+        <h3>{t('Waiting for events…')}</h3>
+        <p>
+          {tct(
+            'Our error robot is waiting to [strike:devour] receive your first event.',
+            {
+              strike: <Strikethrough />,
+            }
+          )}
+        </p>
+        <p>
+          {project && (
+            <Button
+              data-test-id="install-instructions"
+              priority="primary"
+              to={`/${org.slug}/${project.slug}/getting-started/${
+                project.platform || ''
+              }`}
+            >
+              {t('Installation Instructions')}
+            </Button>
+          )}
+        </p>
+        {sampleLink}
+      </MessageContainer>
+    </ErrorRobotWrapper>
+  );
 }
 
-export {ErrorRobot};
+export default ErrorRobot;
 
-export default withApi(ErrorRobot);
-
-const ErrorRobotWrapper = styled('div')<{gradient: boolean}>`
+const ErrorRobotWrapper = styled('div')`
   display: flex;
   justify-content: center;
-  font-size: ${p => p.theme.fontSizeExtraLarge};
-  box-shadow: inset 0 1px 3px rgba(0, 0, 0, 0.08);
+  font-size: ${p => p.theme.fontSizeLarge};
   border-radius: 0 0 3px 3px;
   padding: 40px ${space(3)} ${space(3)};
   min-height: 260px;
