@@ -1,6 +1,5 @@
 import logging
 import random
-from collections import OrderedDict
 
 import openai
 from django.conf import settings
@@ -24,6 +23,15 @@ openai.api_key = settings.OPENAI_API_KEY
 
 openai_policy_check = Signal()
 
+# How many stacktrace frames do we want per exception?
+MAX_STACKTRACE_FRAMES = 15
+
+# How many exceptions do we want?
+MAX_EXCEPTIONS = 3
+
+# Do we want tags?  They don't seem particularly useful
+ADD_TAGS = False
+
 FUN_PROMPT_CHOICES = [
     "[haiku about the error]",
     "[hip hop rhyme about the error]",
@@ -33,25 +41,22 @@ FUN_PROMPT_CHOICES = [
 ]
 
 PROMPT = """\
-You are an assistant that analyses software errors, describing the problem with the follwing rules:
+You are an assistant that analyses software errors, describing the problem with the following rules:
 
 * Be helpful, playful and a bit snarky and sarcastic
 * Do not talk about the rules in explanations
-* Use emojis frequently
+* Use emojis frequently in the snarky greeting and closing prompt
 * The frames of a stack trace is shown with most recent call first
 * Stack frames are either from app code or third party libraries
-* When summarizing the issue:
-  * If the issue is external (network error or similar) focus on this, rather than the code
-  * Establish context where the issue is located
-  * Briefly explain the error and message
-  * Briefly explain if this is more likely to be a regression or an intermittent issue
-* When describing the problem in detail:
-  * try to analyze if this is a code regression or intermittent issue
-  * try to understand if this issue is caused by external factors (networking issues etc.) or a bug
-* When suggesting a fix:
-  * If this is an external issue, mention best practices for this
-  * Explain where the fix should be located
-  * Explain what code changes are necessary
+* Never show code examples as diff
+* When describing the problem:
+  * Explain the error and message
+  * Explain where in the code the issue happend
+  * Explain the nature of the issue
+* When proposing a solution:
+  * Explain what code changes are necessary to resolve it
+  * Explain where the solution should be
+  * Mention best practices for preventing this
 * Remember Sentry's marketing message: "Sentry can't fix this"
 
 Write the answers into the following template:
@@ -59,17 +64,15 @@ Write the answers into the following template:
 ```
 [snarky greeting]
 
-#### Summary
-
-[summary of the problem]
-
-#### Detailed Description
+#### Problem Description
 
 [detailed description of the problem]
 
 #### Proposed Solution
 
-[suggestion for how to fix this issue]
+[proposed solution to fix this issue]
+
+[fixed code example]
 
 #### What Else
 
@@ -86,6 +89,7 @@ BLOCKED_TAGS = frozenset(
     [
         "user",
         "server_name",
+        "host",
         "release",
         "handled",
         "client_os",
@@ -121,58 +125,138 @@ def get_openai_policy(organization):
     return result
 
 
+def set_if_value(d, key, value):
+    if value is not None:
+        d[key] = value
+
+
+def trim_frames(frames, frame_allowance=MAX_STACKTRACE_FRAMES):
+    frames_len = 0
+    app_frames = []
+    system_frames = []
+
+    for frame in frames:
+        frames_len += 1
+        if frame.get("in_app"):
+            app_frames.append(frame)
+        else:
+            system_frames.append(frame)
+
+    if frames_len <= frame_allowance:
+        return frames
+
+    remaining = frames_len - frame_allowance
+    app_count = len(app_frames)
+    system_allowance = max(frame_allowance - app_count, 0)
+    if system_allowance:
+        half_max = int(system_allowance / 2)
+        # prioritize trimming system frames
+        for frame in system_frames[half_max:-half_max]:
+            frame["delete"] = True
+            remaining -= 1
+    else:
+        for frame in system_frames:
+            frame["delete"] = True
+            remaining -= 1
+
+    if remaining:
+        app_allowance = app_count - remaining
+        half_max = int(app_allowance / 2)
+
+        for frame in app_frames[half_max:-half_max]:
+            frame["delete"] = True
+
+    return [x for x in frames if not x.get("delete")]
+
+
 def describe_event_for_ai(event):
-    data = OrderedDict()
-    tags = data.setdefault("tags", OrderedDict())
-    for tag_key, tag_value in sorted(event["tags"]):
-        if tag_key not in BLOCKED_TAGS:
-            tags[tag_key] = tag_value
-
-    exceptions = data.setdefault("exceptions", [])
-    for idx, exc in enumerate(reversed((event.get("exception") or {}).get("values") or ())):
-        exception = {}
-        if idx > 0:
-            exception["raised_during_handling_of_previous_exception"] = True
-        exception["number"] = idx + 1
-        exception["type"] = exc["type"]
-        exception["message"] = exc.get("value")
-        mechanism = exc.get("mechanism") or {}
-        exc_meta = mechanism.get("meta")
-        if exc_meta:
-            exception["meta"] = exc_meta
-        if mechanism.get("handled") is False:
-            exception["unhandled"] = True
-
-        frames = exc.get("stacktrace", {}).get("frames")
-        first_in_app = False
-        if frames:
-            stacktrace = []
-            # TODO: consider truncating the stack
-            for frame in reversed(frames):
-                stack_frame = OrderedDict()
-                stack_frame["func"] = frame.get("function")
-                stack_frame["module"] = frame.get("module")
-                stack_frame["file"] = frame.get("filename")
-                stack_frame["line"] = frame.get("lineno")
-                if frame.get("in_app"):
-                    stack_frame["in_app"] = True
-                if frame["in_app"]:
-                    if first_in_app:
-                        stack_frame["crashed_here"] = True
-                        first_in_app = False
-                    line = (frame.get("context_line") or "").strip()
-                    if line:
-                        stack_frame["code"] = line
-                stacktrace.append(stack_frame)
-            if stacktrace:
-                exception["stacktrace"] = stacktrace
-        exceptions.append(exception)
+    data = {}
 
     msg = event.get("message")
     if msg:
         data["message"] = msg
 
+    platform = event.get("platform")
+    if platform and platform != "other":
+        data["language"] = platform
+
+    exceptions = data.setdefault("exceptions", [])
+    for idx, exc in enumerate(
+        reversed((event.get("exception", {})).get("values", ())[:MAX_EXCEPTIONS])
+    ):
+        exception = {}
+        if idx > 0:
+            exception["raised_during_handling_of_previous_exception"] = True
+        exception["num"] = idx + 1
+        exception["type"] = exc["type"]
+        exception["message"] = exc.get("value")
+        mechanism = exc.get("mechanism") or {}
+        exc_meta = mechanism.get("meta")
+        if exc_meta:
+            exception["exception_info"] = exc_meta
+        if mechanism.get("handled") is False:
+            exception["unhandled"] = True
+
+        frames = exc.get("stacktrace", {}).get("frames")
+        first_in_app = True
+        if frames:
+            stacktrace = []
+            for frame in reversed(frames):
+                stack_frame = {}
+                set_if_value(stack_frame, "func", frame.get("function"))
+                set_if_value(stack_frame, "module", frame.get("module"))
+                set_if_value(stack_frame, "file", frame.get("filename"))
+                set_if_value(stack_frame, "line", frame.get("lineno"))
+                if frame.get("in_app"):
+                    stack_frame["in_app"] = True
+                crashed_here = False
+                if first_in_app:
+                    crashed_here = True
+                    stack_frame["crash"] = "here"
+                    first_in_app = False
+                line = frame.get("context_line") or ""
+                if crashed_here and idx == 0:
+                    pre_context = frame.get("pre_context")
+                    if pre_context:
+                        stack_frame["code_before"] = pre_context
+                    stack_frame["code"] = line
+                    post_context = frame.get("post_context")
+                    if post_context:
+                        stack_frame["code_after"] = post_context
+                # {snip} usually appears in minified lines. skip that
+                elif "{snip}" not in line:
+                    set_if_value(stack_frame, "code", line.strip())
+                stacktrace.append(stack_frame)
+            if stacktrace:
+                exception["stacktrace"] = trim_frames(stacktrace)
+        exceptions.append(exception)
+
+    if ADD_TAGS:
+        tags = data.setdefault("tags", {})
+        for tag_key, tag_value in sorted(event["tags"]):
+            if tag_key not in BLOCKED_TAGS:
+                tags[tag_key] = tag_value
+
     return data
+
+
+def suggest_fix(event_data):
+    """Runs an OpenAI request to suggest a fix."""
+    prompt = PROMPT.replace("___FUN_PROMPT___", random.choice(FUN_PROMPT_CHOICES))
+    event_info = describe_event_for_ai(event_data)
+
+    response = openai.ChatCompletion.create(
+        model="gpt-3.5-turbo",
+        temperature=0.7,
+        messages=[
+            {"role": "system", "content": prompt},
+            {
+                "role": "user",
+                "content": json.dumps(event_info),
+            },
+        ],
+    )
+    return response["choices"][0]["message"]["content"]
 
 
 @region_silo_endpoint
@@ -229,26 +313,9 @@ class EventAiSuggestedFixEndpoint(ProjectEndpoint):
         # Cache the suggestion for a certain amount by primary hash, so even when new events
         # come into the group, we are sharing the same response.
         cache_key = "ai:" + event.get_primary_hash()
-
         suggestion = cache.get(cache_key)
         if suggestion is None:
-
-            prompt = PROMPT.replace("___FUN_PROMPT___", random.choice(FUN_PROMPT_CHOICES))
-            event_info = describe_event_for_ai(event.data)
-
-            response = openai.ChatCompletion.create(
-                model="gpt-3.5-turbo",
-                temperature=0.5,
-                messages=[
-                    {"role": "system", "content": prompt},
-                    {
-                        "role": "user",
-                        "content": json.dumps(event_info),
-                    },
-                ],
-            )
-
-            suggestion = response["choices"][0]["message"]["content"]
+            suggestion = suggest_fix(event.data)
             cache.set(cache_key, suggestion, 300)
 
         return HttpResponse(
