@@ -13,7 +13,7 @@ import {
   generateDocKeys,
   isPlatformSupported,
 } from 'sentry/components/replaysOnboarding/utils';
-import OnboardingStep from 'sentry/components/sidebar/onboardingStep';
+import {DocumentationWrapper} from 'sentry/components/sidebar/onboardingStep';
 import SidebarPanel from 'sentry/components/sidebar/sidebarPanel';
 import {CommonSidebarProps, SidebarPanelKey} from 'sentry/components/sidebar/types';
 import {replayPlatforms} from 'sentry/data/platformCategories';
@@ -184,12 +184,7 @@ function OnboardingContent({currentProject}: {currentProject: Project}) {
         }
         return (
           <div key={index}>
-            <OnboardingStep
-              docContent={docContents[docKey]}
-              docKey={docKey}
-              prefix="replay"
-              project={currentProject}
-            />
+            <OnboardingStepV2 step={index + 1} content={docContents[docKey]} />
             {footer}
           </div>
         );
@@ -198,8 +193,47 @@ function OnboardingContent({currentProject}: {currentProject: Project}) {
   );
 }
 
+// TODO: we'll have to move this into a folder for common consumption w/ Profiling, Performance etc.
+interface OnboardingStepV2Props {
+  content: string;
+  step: number;
+}
+
+function OnboardingStepV2({step, content}: OnboardingStepV2Props) {
+  return (
+    <OnboardingStepContainer>
+      <div>
+        <TaskStepNumber>{step}</TaskStepNumber>
+      </div>
+      <div>
+        <DocumentationWrapper dangerouslySetInnerHTML={{__html: content}} />
+      </div>
+    </OnboardingStepContainer>
+  );
+}
+
+const OnboardingStepContainer = styled('div')`
+  display: flex;
+  & > :last-child {
+    overflow: hidden;
+  }
+`;
+
+const TaskStepNumber = styled('div')`
+  display: flex;
+  margin-right: ${space(1.5)};
+  background-color: ${p => p.theme.yellow300};
+  border-radius: 50%;
+  font-weight: bold;
+  height: ${space(4)};
+  width: ${space(4)};
+  justify-content: center;
+  align-items: center;
+`;
+
 const TaskSidebarPanel = styled(SidebarPanel)`
-  width: 450px;
+  width: 600px;
+  max-width: 100%;
 `;
 
 const TopRightBackgroundImage = styled('img')`
