@@ -1,6 +1,6 @@
 import {render, screen} from 'sentry-test/reactTestingLibrary';
 
-import {useQuery} from 'sentry/utils/queryClient';
+import {useApiQuery} from 'sentry/utils/queryClient';
 import RequestError from 'sentry/utils/requestError/requestError';
 import * as useApi from 'sentry/utils/useApi';
 
@@ -21,7 +21,7 @@ describe('queryClient', function () {
       });
 
       const TestComponent = () => {
-        const {data} = useQuery<ResponseData>(['/some/test/path/'], {staleTime: 0});
+        const {data} = useApiQuery<ResponseData>(['/some/test/path/'], {staleTime: 0});
 
         if (!data) {
           return null;
@@ -44,7 +44,7 @@ describe('queryClient', function () {
       });
 
       const TestComponent = () => {
-        const {data} = useQuery<ResponseData>(
+        const {data} = useApiQuery<ResponseData>(
           ['/some/test/path/', {query: {filter: 'red'}}],
           {staleTime: 0}
         );
@@ -68,7 +68,7 @@ describe('queryClient', function () {
 
     it('can fetch with custom query function', async function () {
       const TestComponent = () => {
-        const {data} = useQuery<ResponseData>(['some-key'], () => ({value: 5}));
+        const {data} = useApiQuery<ResponseData>(['some-key'], () => ({value: 5}));
 
         if (!data) {
           return null;
@@ -93,7 +93,7 @@ describe('queryClient', function () {
       jest.spyOn(api, 'requestPromise').mockRejectedValue(requestError);
 
       const TestComponent = () => {
-        const {isError, error} = useQuery<ResponseData>(['/some/test/path'], {
+        const {isError, error} = useApiQuery<ResponseData>(['/some/test/path'], {
           staleTime: 0,
         });
 
