@@ -68,7 +68,7 @@ function isQueryFn<TQueryFnData, TError, TData>(
  *   {staleTime: 0}
  * );
  */
-function useQuery<TQueryFnData, TError = RequestError, TData = TQueryFnData>(
+function useApiQuery<TQueryFnData, TError = RequestError, TData = TQueryFnData>(
   queryKey: QueryKey,
   queryOptions: UseQueryOptions<TQueryFnData, TError, TData>
 ): reactQuery.UseQueryResult<TData, TError>;
@@ -81,12 +81,12 @@ function useQuery<TQueryFnData, TError = RequestError, TData = TQueryFnData>(
  *   {staleTime: 0}
  * )
  */
-function useQuery<TQueryFnData, TError = RequestError, TData = TQueryFnData>(
+function useApiQuery<TQueryFnData, TError = RequestError, TData = TQueryFnData>(
   queryKey: QueryKey,
   queryFn: reactQuery.QueryFunction<TQueryFnData, QueryKey>,
   queryOptions?: UseQueryOptions<TQueryFnData, TError, TData>
 ): reactQuery.UseQueryResult<TData, TError>;
-function useQuery<TQueryFnData, TError = RequestError, TData = TQueryFnData>(
+function useApiQuery<TQueryFnData, TError = RequestError, TData = TQueryFnData>(
   queryKey: QueryKey,
   queryFnOrQueryOptions:
     | reactQuery.QueryFunction<TQueryFnData, QueryKey>
@@ -121,5 +121,7 @@ function useQuery<TQueryFnData, TError = RequestError, TData = TQueryFnData>(
 // eslint-disable-next-line import/export
 export * from '@tanstack/react-query';
 
+const useQuery = useApiQuery;
+
 // eslint-disable-next-line import/export
-export {DEFAULT_QUERY_CLIENT_CONFIG, useQuery, UseQueryOptions};
+export {DEFAULT_QUERY_CLIENT_CONFIG, useApiQuery, useQuery, UseQueryOptions};
