@@ -8,7 +8,6 @@ import EmptyMessage from 'sentry/components/emptyMessage';
 import FeatureBadge from 'sentry/components/featureBadge';
 import {feedbackClient} from 'sentry/components/featureFeedback/feedbackModal';
 import LoadingError from 'sentry/components/loadingError';
-import LoadingIndicator from 'sentry/components/loadingIndicator';
 import {Panel, PanelBody, PanelFooter, PanelHeader} from 'sentry/components/panels';
 import {IconFile, IconFlag, IconHappy, IconMeh, IconSad} from 'sentry/icons';
 import {IconChevron} from 'sentry/icons/iconChevron';
@@ -16,7 +15,7 @@ import {t} from 'sentry/locale';
 import ConfigStore from 'sentry/stores/configStore';
 import {space} from 'sentry/styles/space';
 import marked from 'sentry/utils/marked';
-import {useQuery} from 'sentry/utils/queryClient';
+import {useApiQuery} from 'sentry/utils/queryClient';
 import useOrganization from 'sentry/utils/useOrganization';
 import useRouter from 'sentry/utils/useRouter';
 import {AiLoaderMessage} from 'sentry/views/issueDetails/openAIFixSuggestion/aiLoaderMessage';
@@ -79,7 +78,7 @@ export function OpenAIFixSuggestionPanel({eventID, projectSlug}: Props) {
     isError: dataIsError,
     refetch: dataRefetch,
     error,
-  } = useQuery<{suggestion: string}>(
+  } = useApiQuery<{suggestion: string}>(
     [
       `/projects/${organization.slug}/${projectSlug}/events/${eventID}/ai-fix-suggest/`,
       {query: {consent: individualConsent ? 'yes' : undefined}},
@@ -179,16 +178,12 @@ export function OpenAIFixSuggestionPanel({eventID, projectSlug}: Props) {
       </FixSuggestionPanelHeader>
       {expandedSuggestedFix && (
         <Fragment>
-          <StyledPanelBody withPadding whiteBg={dataIsLoading && individualConsent}>
+          <StyledPanelBody withPadding whiteBg={dataIsLoading}>
             {dataIsLoading ? (
-              !individualConsent ? (
-                <LoadingIndicator />
-              ) : (
-                <AiLoaderWrapper>
-                  <div className="ai-loader" />
-                  <AiLoaderMessage />
-                </AiLoaderWrapper>
-              )
+              <AiLoaderWrapper>
+                <div className="ai-loader" />
+                <AiLoaderMessage />
+              </AiLoaderWrapper>
             ) : dataIsError ? (
               PolicyErrorState ? (
                 PolicyErrorState
