@@ -13,7 +13,12 @@ import {
   TagValue,
 } from 'sentry/types';
 import {isAggregateField} from 'sentry/utils/discover/fields';
-import {FieldKind, getFieldDefinition, REPLAY_FIELDS} from 'sentry/utils/fields';
+import {
+  FieldKind,
+  getFieldDefinition,
+  REPLAY_CLICK_FIELDS,
+  REPLAY_FIELDS,
+} from 'sentry/utils/fields';
 import useApi from 'sentry/utils/useApi';
 import useTags from 'sentry/utils/useTags';
 
@@ -37,15 +42,16 @@ function fieldDefinitionsToTagCollection(fieldKeys: string[]): TagCollection {
       {
         key,
         name: key,
-        kind: getReplayFieldDefinition(key)?.kind,
+        ...getReplayFieldDefinition(key),
       },
     ])
   );
 }
 
 const REPLAY_FIELDS_AS_TAGS = fieldDefinitionsToTagCollection(REPLAY_FIELDS);
+const REPLAY_CLICK_FIELDS_AS_TAGS = fieldDefinitionsToTagCollection(REPLAY_CLICK_FIELDS);
 
-function getSupportedTags(supportedTags: TagCollection) {
+function getSupportedTags(supportedTags: TagCollection, organization: Organization) {
   return {
     ...Object.fromEntries(
       Object.keys(supportedTags).map(key => [
@@ -56,6 +62,9 @@ function getSupportedTags(supportedTags: TagCollection) {
         },
       ])
     ),
+    ...(organization && organization.features.includes('session-replay-dom-search')
+      ? REPLAY_CLICK_FIELDS_AS_TAGS
+      : {}),
     ...REPLAY_FIELDS_AS_TAGS,
   };
 }
@@ -103,7 +112,7 @@ function ReplaySearchBar(props: Props) {
     <SmartSearchBar
       {...props}
       onGetTagValues={getTagValues}
-      supportedTags={getSupportedTags(tags)}
+      supportedTags={getSupportedTags(tags, organization)}
       placeholder={t('Search for users, duration, count_errors, and more')}
       prepareQuery={prepareQuery}
       maxQueryLength={MAX_QUERY_LENGTH}
